git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc/terrain@24875 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/snow/persist/standard/snowldr.cpp
@@ -366,11 +366,11 @@ bool csSnowSaver::WriteDown (iBase* obj, iDocumentNode* parent,
     synldr->WriteBool(paramsNode, "lighting", snowstate->GetLighting(), true);
 
     //Writedown Number tag
-    int number = snowstate->GetParticleCount();
+    size_t number = snowstate->GetParticleCount();
     csRef<iDocumentNode> numberNode = paramsNode->CreateNodeBefore(CS_NODE_ELEMENT, 0);
     numberNode->SetValue("number");
     csRef<iDocumentNode> numberValueNode = numberNode->CreateNodeBefore(CS_NODE_TEXT, 0);
-    numberValueNode->SetValueAsInt(number);
+    numberValueNode->SetValueAsInt((int)number);
   }
 
   paramsNode=0;
